% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.r
\name{check_data}
\alias{check_data}
\title{Data Checking Function}
\usage{
check_data(data)
}
\arguments{
\item{data}{A vector, class 'circular' is recommended but not required}
}
\description{
Make sure data is in the right format.  Datasets are coerced into class 'circular'
of type = angles, units = radians, and modulo = 2pi.  It is recommended to set
these atributes ahead of time.
}
\examples{
testdata = circular::rvonmises(100, mu = circular::circular(pi), kappa = 3,
   control.circular = list(units = "degrees"))
check_data(testdata)
}
\keyword{checking}
\keyword{data}
