% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1.r
\name{M1}
\alias{M1}
\title{Model M1 MLE function}
\usage{
M1(data)
}
\arguments{
\item{data}{A vector of class 'circular'}
}
\value{
A list with the elements:

$lik: The negative log likelihood of data for model M1
}
\description{
Run Maximum likelihood estimation for model M1.
}
\examples{
testdata = circular::rvonmises(100, mu = circular::circular(pi), kappa = 3)
M1(testdata)
}
\keyword{M1}
