% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pycke.r
\name{pycke_test}
\alias{pycke_test}
\title{Pycke test function}
\usage{
pycke_test(data, iter = 9999)
}
\arguments{
\item{data}{A vector of class 'circular', or numeric vector of 
angles measured in radians}

\item{iter}{The number of bootstrap replicates to perform in order 
to estimate the p-value of the test. (default = 9999)}
}
\value{
A numeric vector of the test statistic (T) and associated p-value
}
\description{
Perform the Pycke test.
}
\examples{
testdata = circular::rvonmises(20, mu = circular::circular(pi), kappa = 3)
pycke_test(testdata, iter = 999)
}
\keyword{Pycke}
\keyword{test}
